test: split group auth tests
